rename game to game AI
